1) fix warnings: -Werror 2) 修复部分用例对路径有依赖的问题。 3) 适配gcc 9.3.1部分特性(return rvalue)

+++ core/cms/common/impl/ResourceConsumptionRecord.cpp
@@ -19,7 +19,7 @@ std::string ResourceConsumption::str() const {
 
 /// ////////////////////////////////////////////////////////////////////////////////////////////////////////////////////
 ///
-constexpr const int ResourceConsumptionRecorder::topN;
+constexpr const size_t ResourceConsumptionRecorder::topN;
 ResourceConsumptionRecorder::ResourceConsumptionRecorder() = default;
 
 // void ResourceConsumptionRecorder::EnableRecording(bool enable) {
